refactor ticket importer and update tickets to ticket
