+++ app/models/ticket.rb
